Got Deb's script working to output the 3 main figures
Updated SST_rain_comparison_bin_average.m so that it would run succesfully without references to obsolete/nonexistent functions. Outputs the 3 main plots we are taking as a starting point.

Updated importing lines for SST_rain_comparison_ID_rainpeaks, but could not find the Revelle met data needed for this script - likely on cw2 on Ragdoll, to which I do not have samba access yet.